pyproject.toml
[tool.poetry]
name = "forward-telegram-bot"
version = "0.1.0"
description = ""
authors = ["riguima <richard.alexsander.guima@gmail.com>"]
readme = "README.md"
packages = [{include = "forward_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.16"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
test = "pytest -s -x --cov=forward_telegram_bot --cov-report=html:coverage -vv"
